Replaces uses of doubles with floats
Looks like the KallistiOS Dreamcast SDK disables double support
by default: https://github.com/KallistiOS/KallistiOS/blob/495e77fd60d5b09a1ad52a26cd4a7e73cd0d9d51/environ_dreamcast.sh#L16

We don't really need doubles in this code.

The one place where we might have needed them is the SMK video
decoder, handled in a separate PR.

+++ Source/controls/touch/gamepad.cpp
@@ -12,40 +12,40 @@ VirtualGamepad VirtualGamepadState;
 
 namespace {
 
-constexpr double Pi = 3.141592653589793;
+constexpr float Pi = 3.141592653589793F;
 
-int roundToInt(double value)
+int roundToInt(float value)
 {
 	return static_cast<int>(round(value));
 }
 
-constexpr bool PointsUp(double angle)
+constexpr bool PointsUp(float angle)
 {
-	constexpr double UpAngle = Pi / 2;
-	constexpr double MinAngle = UpAngle - 3 * Pi / 8;
-	constexpr double MaxAngle = UpAngle + 3 * Pi / 8;
+	constexpr float UpAngle = Pi / 2;
+	constexpr float MinAngle = UpAngle - 3 * Pi / 8;
+	constexpr float MaxAngle = UpAngle + 3 * Pi / 8;
 	return MinAngle <= angle && angle <= MaxAngle;
 }
 
-constexpr bool PointsDown(double angle)
+constexpr bool PointsDown(float angle)
 {
-	constexpr double DownAngle = -Pi / 2;
-	constexpr double MinAngle = DownAngle - 3 * Pi / 8;
-	constexpr double MaxAngle = DownAngle + 3 * Pi / 8;
+	constexpr float DownAngle = -Pi / 2;
+	constexpr float MinAngle = DownAngle - 3 * Pi / 8;
+	constexpr float MaxAngle = DownAngle + 3 * Pi / 8;
 	return MinAngle <= angle && angle <= MaxAngle;
 }
 
-constexpr bool PointsLeft(double angle)
+constexpr bool PointsLeft(float angle)
 {
-	constexpr double MaxAngle = Pi - 3 * Pi / 8;
-	constexpr double MinAngle = -Pi + 3 * Pi / 8;
+	constexpr float MaxAngle = Pi - 3 * Pi / 8;
+	constexpr float MinAngle = -Pi + 3 * Pi / 8;
 	return !(MinAngle < angle && angle < MaxAngle);
 }
 
-constexpr bool PointsRight(double angle)
+constexpr bool PointsRight(float angle)
 {
-	constexpr double MinAngle = -3 * Pi / 8;
-	constexpr double MaxAngle = 3 * Pi / 8;
+	constexpr float MinAngle = -3 * Pi / 8;
+	constexpr float MaxAngle = 3 * Pi / 8;
 	return MinAngle <= angle && angle <= MaxAngle;
 }
 
@@ -227,14 +227,14 @@ void VirtualDirectionPad::UpdatePosition(Point touchCoordinates)
 	if (!area.contains(position)) {
 		int x = diff.deltaX;
 		int y = diff.deltaY;
-		double dist = sqrt(x * x + y * y);
+		float dist = sqrt(x * x + y * y);
 		x = roundToInt(x * area.radius / dist);
 		y = roundToInt(y * area.radius / dist);
 		position.x = area.position.x + x;
 		position.y = area.position.y + y;
 	}
 
-	double angle = atan2(-diff.deltaY, diff.deltaX);
+	float angle = atan2(-diff.deltaY, diff.deltaX);
 
 	isUpPressed = PointsUp(angle);
 	isDownPressed = PointsDown(angle);
